add gen docs file, refactoring
